[Refactor] Move failure logging hooks to service layer
