+++ be/src/service/daemon.cpp
@@ -74,9 +74,9 @@
 #include "platform/platform_metrics.h"
 #include "platform/user_function_cache.h"
 #include "runtime/exec_env.h"
-#include "runtime/runtime_logging.h"
 #include "runtime/runtime_metrics.h"
 #include "service/backend_metrics_initializer.h"
+#include "service/failure_handler.h"
 #include "service/mem_hook.h"
 #include "storage/storage_engine.h"
 #include "storage/storage_metrics.h"
